start work on tab bar
