+++ Podcast/BookmarkViewController.swift
@@ -21,7 +21,7 @@ class BookmarkViewController: ViewController, EmptyStateTableViewDelegate, UITab
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .paleGrey
-        title = "Saved for Later"
+        navigationItem.title = "Saved for Later"
     
         //tableview.
         bookmarkTableView = EmptyStateTableView(frame: view.frame, type: .bookmarks, isRefreshable: true)
@@ -132,7 +132,7 @@ class BookmarkViewController: ViewController, EmptyStateTableViewDelegate, UITab
     
     func didPressEmptyStateViewActionItem() {
         guard let appDelegate = UIApplication.shared.delegate as? AppDelegate, let tabBarController = appDelegate.tabBarController else { return }
-        tabBarController.programmaticallyPressTabBarButton(atIndex: System.discoverTab) //discover index
+        tabBarController.selectedIndex = System.discoverTab
     }
     
     //MARK
