+++ Podcast/UserDetailViewController.swift
@@ -247,7 +247,7 @@ extension UserDetailViewController: UITableViewDelegate, UITableViewDataSource {
             navigationController?.pushViewController(episodeDetailViewController, animated: true)
         } else {
             guard let appDelegate = UIApplication.shared.delegate as? AppDelegate, let tabBarController = appDelegate.tabBarController else { return }
-            tabBarController.programmaticallyPressTabBarButton(atIndex: System.discoverTab)
+            tabBarController.selectedIndex = System.discoverTab
         }
     }
     
@@ -291,7 +291,7 @@ extension UserDetailViewController: UICollectionViewDataSource, UICollectionView
             navigationController?.pushViewController(seriesDetailViewController, animated: true)
         } else {
             guard let appDelegate = UIApplication.shared.delegate as? AppDelegate, let tabBarController = appDelegate.tabBarController else { return }
-            tabBarController.programmaticallyPressTabBarButton(atIndex: System.discoverTab)
+            tabBarController.selectedIndex = System.discoverTab
         }
     }
